♻️ refactor:codeFactory에 맞게 리팩토링'
